Add files via upload
The data set called the Civil Rights Data Collection. It contains information on educational achievement and opportunities in the U.S., broken down by race and school. For example, it records the racial composition of the students enrolled in advanced classes at each school. Each row represents a school, while each column records an indicator of academic achievement.

This code will be using a subset of the data that only contains 1 out of every 100 rows in the original data set.
4 data files are stored in the data folder
3 code files
1 findings.txt file